Replace my old vx32 subdirectory (from the USENIX FAST paper) with the final vx32 source tree from after the USENIX vx32 paper, excluding the large external source trees (spec06, gcc, binutils).
This commit represents a merge with an appropriately pruned
history derived from the USENIX vx32 mercurial repository,
to ensure that we have the history in this repository in some form.
(The complete USENIX vx32 paper repository may still be found
in its original mercurial form on korz:~git/repositories/old/2008-vx32.hg)

Note that the old vx32 source tree that this commit replaces
had the (bare) beginnings of a PowerPC translator,
and an almost-working instruction interpreter,
in case it ever proves useful to resurrect them.