schnorr: signature slice of bytes compat. with eddsa
This commit adds a sign/ package with the schnorr signature and verification
routines inside. These routines have been modified to take a slice of bytes as a
signature instead of a struct and also to make the signature EdDSA compatible by
hashing the public key with in the hash function. The hash function here is
given by the suite; with ed25519, it is SHA256.