Added validHostname as isolated function. (#179)
* Added validHostname as isolated function.
Added few (temporary) tests
It has to be integrated with the rest

* Formatting issues solved

* Modified the NetworkAddress() function, so that it always returns an IP address (even if the Address is defined with a hostname).
validHostname function -> partially integrated in the Valid() function. A refactor of the Valid() function might be needed (for instance: the check ip == "localhost" is already handled by validHostname

* TLD do not allow hyphens/numbers (but a minor change might be needed in order to support IDNs)
Small change in the Valid() function
Added few tests

* TLD do not allow hyphens/numbers (but a minor change might be needed in order to support IDNs)
Small change in the Valid() function
Added few tests

* Corrected formatting and linting.

* Removed useless comment

* Added:
 IsHostname()
 Resolve() -> returned the IP address associated to the hostname
 NetworkAddressResolved() -> returns the network address, but with the hostname resolved
Still need to handle IPv6 -> does it make sense for the Host() method to return the host part inside brackets? The "net" library handles brackets automatically when splitting and joining addresses/ports. Also, LookupHost does not return the IPv6 address inside brackets. Finally, ParseIP requires the IPv6 address without brackets.
Valid() function -> is it really ok that only the structure of the address is checked?