Changes to make things compile with kyber.master
Added network.DefaultSuite, which would let us change the
entire crypto suite at runtime via environment variables, if we want.

Postponed calculation of Roster aggregate key as part of
attempts to make the cipher suite more configurable. (Was not
really necessary in fact.)

Rationalized use of tSuite and suite, opting for tSuite because it
is a more distinctive string to read and search for.