Convert proof to XOF
Added Reseed method to XOF, which encapsulates the process
of chaining two XOFs in order to work around the limitation that
you cannot write to an XOF after you read from it.

Proof is the first user of this method.

Shuffle depends on proof.Suite, so it needed to be converted to
using an XOF random stream at the same time.