Fixed a bug with mismatch enc/dec
See the new test to see where the problem comes from.
Without the associated fix in blake/blake.go and
keccak/keccak.go, it fails.

The fix is to be certain to always take exactly the same
numbers of key bytes from the XOF as src bytes.

Also: improve docs.