Centralize logic about db path and deletion
Now there's one place where the policy is decided,
at the top of newServer. The policy is passed into
serviceManager, which implments the delete on close.

Before this work, simulations were writing into a tmp
directory. With this change, they now use the environment
variable to write into the simulation directory. This
will be cleaned up as it was before.

Some tests got a bit more complicated, because they are
now piggybacking off of the NewLocalTest/defer CloseAll()
in order to create/delete a directory for their server to use.