Edwards25519 SetBytes must follow the interface definition
SetBytes is defined in kyber.Suite as bytes are big endian
byte order, so things to work like that.

Marshal/UnmarshalBinary is still LE, so that we interoperate with
the rest of the world's Ed25519 implementations.

Also: Remove some dead code I noticed.

Also: Package encoding had circular dependency problems. Fix
with the _test trick.

Fixes #274.