Fix vss marshaling to work with fixed protobuf encoder
Before the BinaryMarshaler detection was not working. When
we fixed the BinaryMarashaler detection, it turned out this
code had always been wrong (recursive loop) but never called.
Now we use the simplest thing possible: protobuf.Encode for
encoding, and the previous code for decoding (but not triggering
the BinaryUnmarshaler detection).