A mechanism for callers to opt-in to a limit on suites
Previously, vartime suites were not even compiled in. We found this to
be difficult to work with when we need to mix vartime and constant
time algorithms (i.e. in Cothority). This is a compromise: by having
callers add one line to their init() function in package main,
they can opt-in to a different implementation that still provides
the same safety as the old implementation.