Feat/broadcast (#18)
* Verification of subgroup element for schnorr sigs

When using a non prime order group over an elliptic curve, a point of
the curve can belong to the correct curve but not necessarily to the correct
prime order subgroup subgroup. That commits additionally ensures that this is
the case with the given point if the functionality allows for it.
The scalar must be checked as well that it is in the right range, but currently
group/mod is used which automatically checks that property.

* Outsource the verification of a DKG packet

DKG packets can now be verified _before_ they are passed to the internal
protocol if the option is specified. That allows to preprocess the packet before
passing them to the application if a given packet needs to be rebroadcasted,
which is the case for broadcast protocol.