Schnorr signature verification with a null public key is an error
Schnorr signatures are only defined for non-zero private keys because the
corresponding public key to a zero private key is the mulplicative identity,
causing the message hash to be excluded from the signature verification.

Do not allow null public keys so that if one is passed in by accident, Verify
returns an error instead of saying that the signature is good, even if the
message was modified.

It is still the caller's responsibility to use Kyber to make
valid secret keys.